final commit

+++ assets/components/basket.js
@@ -27,23 +27,29 @@ $(document).ready(function(){
         let img = this.previousElementSibling.previousElementSibling.firstElementChild.firstElementChild.src;
         let name = this.previousElementSibling.firstElementChild.innerText;
         let price = this.previousElementSibling.firstElementChild.nextElementSibling.innerText;
-        let obj = {
-            img, name, price
-        }
+        let count = 1;
         let basket = JSON.parse(localStorage.getItem("basketLocal"));
-        basket.push(obj);
-        $("#BASKET_COUNT")[0].innerText = basket.length
-        localStorage.setItem("basketLocal", JSON.stringify(basket));
-        // console.log($(".BASKET_COUNT")[0].innerText)
-        basketTemp = localStorage.getItem("basketLocal");
-        if (basketTemp.length == 0) {
-            $("#cart-div__shopping")[0].innerText = "Your shopping cart is empty!"
+        let existed = basket.find((val, index, arr) => {
+           return val.name == name
+        })
+        if (existed) {
+            existed.count++;
+            localStorage.setItem("basketLocal", JSON.stringify(basket));
         }
         else{
-            for(let i = 0; i < basketTemp.length; i++){
-                $("#cart-div__shopping")[0].innerText += JSON.stringify(basketTemp[i]);
-            }
+            let obj = 
+        {
+            img, name, price, count
         }
+        
+        basket.push(obj);
+        localStorage.setItem("basketLocal", JSON.stringify(basket));
+        }
+        
+        $("#BASKET_COUNT")[0].innerText = basket.length
+        
+       
+        
         
     })
 })
